refactor: move LazyLoadingCommand to Command subnamespace
For consistency.
