+++ test/ContainerCommandLoaderTest.php
@@ -10,9 +10,9 @@ declare(strict_types=1);
 
 namespace LaminasTest\Cli;
 
+use Laminas\Cli\Command\LazyLoadingCommand;
 use Laminas\Cli\ContainerCommandLoader;
 use Laminas\Cli\ContainerResolver;
-use Laminas\Cli\LazyLoadingCommand;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Console\Command\Command;
